+++ lib/td_dd/data_structures/bulk_update.ex
@@ -40,8 +40,12 @@ defmodule TdDd.DataStructures.BulkUpdate do
       rows
       |> Enum.with_index()
       |> Enum.map(fn {row, index} -> {row, index + 2} end)
-      |> Enum.map(fn {%{"external_id" => external_id} = row, index} ->
-        {row, DataStructures.get_data_structure_by_external_id(external_id), index}
+      |> Enum.map(fn
+        {%{"external_id" => external_id} = row, index} ->
+          {row, DataStructures.get_data_structure_by_external_id(external_id), index}
+
+        {row, index} ->
+          {row, nil, index}
       end)
       |> Enum.filter(fn {_row, data_structure, _index} -> data_structure end)
       |> Enum.reduce_while([], fn {row, data_structure, index}, acc ->
@@ -52,6 +56,7 @@ defmodule TdDd.DataStructures.BulkUpdate do
       end)
       |> case do
         [_ | _] = contents -> contents
+        [] -> {:error, %{message: :external_id_not_found}}
         errors -> {:error, errors}
       end
     end
@@ -65,7 +70,7 @@ defmodule TdDd.DataStructures.BulkUpdate do
     rows =
       path
       |> Path.expand()
-      |> File.stream!()
+      |> File.stream!([:trim_bom])
       |> Stream.map(&recode/1)
       |> Stream.reject(&(String.trim(&1) == ""))
       |> CSV.decode!(separator: ?;, headers: true)
@@ -184,6 +189,7 @@ defmodule TdDd.DataStructures.BulkUpdate do
       &bulk_update_notes(&1, &2, data_structures, params, user_id, auto_publish)
     )
     |> Repo.transaction()
+    |> on_complete()
   end
 
   defp bulk_update_notes(_repo, _changes_so_far, data_structures, params, user_id, auto_publish) do
@@ -197,7 +203,9 @@ defmodule TdDd.DataStructures.BulkUpdate do
   end
 
   defp update_structure_notes(data_structure, params, user_id, auto_publish) do
-    case StructureNotesWorkflow.create_or_update(data_structure, params, user_id, auto_publish) do
+    opts = [auto_publish: auto_publish, is_bulk_update: true]
+
+    case StructureNotesWorkflow.create_or_update(data_structure, params, user_id, opts) do
       {:ok, structure_note} -> {:ok, structure_note}
       error -> {error, data_structure}
     end
@@ -246,6 +254,14 @@ defmodule TdDd.DataStructures.BulkUpdate do
 
   defp on_complete({:ok, %{updates: updates} = result}) do
     updates
+    |> Map.keys()
+    |> IndexWorker.reindex()
+
+    {:ok, result}
+  end
+
+  defp on_complete({:ok, %{update_notes: update_notes} = result}) do
+    update_notes
     |> Map.keys()
     |> IndexWorker.reindex()